Complete app factory pattern and create run.py entry point
- Moved create_app() function to app/__init__.py
- Moved db and login_manager initialization to package level
- Created run.py as new application entry point
- Deleted old app.py (functionality moved to app/__init__.py)
- Fixed template and static folder paths for app factory pattern
- All features tested and working

Completes #25
